LPS-74966 BatchablePipe and IncreasableEntry should have been in portal-impl and com.liferay.portal.increment should not have been released. Move over classes into internal package and deprecate released ones.
