+++ portal-impl/src/com/liferay/portal/internal/increment/BufferedIncrementAdvice.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.increment;
+package com.liferay.portal.internal.increment;
 
 import com.liferay.portal.kernel.cache.key.CacheKeyGenerator;
 import com.liferay.portal.kernel.cache.key.CacheKeyGeneratorUtil;
@@ -40,9 +40,7 @@ import org.aopalliance.intercept.MethodInvocation;
 /**
  * @author Zsolt Berentey
  * @author Shuyang Zhou
- * @deprecated As of 7.0.0, with no direct replacement
  */
-@Deprecated
 public class BufferedIncrementAdvice
 	extends AnnotationChainableMethodAdvice<BufferedIncrement> {
