+++ portal-impl/src/com/liferay/portal/internal/increment/IncreasableEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.concurrent;
+package com.liferay.portal.internal.increment;
 
 import com.liferay.portal.kernel.util.HashUtil;
 import com.liferay.portal.kernel.util.StringBundler;
@@ -21,9 +21,7 @@ import java.util.Objects;
 
 /**
  * @author Shuyang Zhou
- * @deprecated As of 7.0.0, with no direct replacement
  */
-@Deprecated
 public abstract class IncreasableEntry<K, V> {
 
 	public IncreasableEntry(K key, V value) {
